restructuring of analysis code: - run all genMatch options in same job - apply genMatching to leptons and hadronic taus - apply fake background estimation consistently to leptons and hadronic taus

+++ python/sbatchManager.py
@@ -76,8 +76,8 @@ class sbatchManager:
   def __init__(self):
     self.workingDir = None
     self.logFileDir = None
-    self.queue = "short"
-    ##self.queue = "long"
+    ##self.queue = "short"
+    self.queue = "long"
     self.command_submit = "sbatch"
     self.command_poll = "squeue"
     self.poll_interval = 30
